fix text eliding for arabic and syriac
Arabic and Syriac are connected scripts where the letter shape changes
depending on the context. Text eliding should not affect that letter
shape if the truncation happens in the middle of a word.

The patch ensures that by adding a Unicode ZWJ character between the
text and the eliding in case the character would connect in the full
string.

Reviewed-by: Simon Hausmann