Added QMAKE_PREFIX_SHLIB and QMAKE_SYMBIAN_SHLIB support.
The first enables you to use a prefix in front of the library. It is
by default enabled for all platforms which previously used "lib", and
on Symbian it is set to the empty string.

The second variable enables Symbian style linking, where you link
to a stub .lib file, but the actual runtime library is .dll.