Remove implicit margin on bulleted lists when exporting to HTML
When we export a bulleted list to HTML, we add a meta-block around the
list items to set the bullet style of the list (<ul> or <ol>.) This
block will implicitly have a margin, even if the list itself does not
in the internal document. The result is that exporting a list to HTML
and then reloading it will give a different apperance, because the 0
margin of the original was not saved. The fix is to always set the
meta-block's margin to 0, since it is only there to set the bullet
style and should not affect the layout at all.

Task-number: 201228
Reviewed-by: Simon Hausmann