Implement symbol hiding for JSC's JIT functions.
These functions are implemented directly in assembly, so they need the
proper directives to enable/disable visibility. On ELF systems, it's
.hidden, whereas on Mach-O systems (Mac) it's .private_extern. On
Windows, it's not necessary since you have to explicitly export. I
also implemented the AIX idiom, though it's unlikely anyone will
implement AIX/POWER JIT.

That leaves only HP-UX on PA-RISC unimplemented, from the platforms
that Qt supports. It's also unlikely that we'll imlpement JIT for it.

Reviewed-by: Kent Hansen