Support setting font size in shorthand font properties
The fontSizeAdjustment defaulted to 0 which means "medium" in internal
semantics. This will override any font size you set in the short-hand.
In other locations, fontSizeAdjustment defaults to -255 which has no
meaning attached. To allow setting the font size in short-hand (as in
"font: 20px Arial"), we can't default to a specific adjustment.

Two tests are added: The first verifies the case that already worked,
where you specify the font size using the "font-size" property. The
other verifies the short-hand case and would previously fail.

Task-number: 207189
Reviewed-by: Simon Hausmann