Createpackage support for signing SIS with multiple certificates.
This commit extends createpackage with capability to sign the SIS
file with several certificates. The certificates need to be listed
in text file. The file can have several certificates, each specified in
separate line. The certificate, key and passphrase in line must be ';'
separated. Lines starting with '#' are treated as a comments. Also
empty lines in the file are ignored. The paths in certificate info file
can be absolute or relative to the given file.

Example syntax for certificateinfo file (mycerts.txt):

    # This is comment line, also the empty lines are ignored
    rd.cer;rd-key.pem
    .\cert\mycert.cer;.\cert\mykey.key;yourpassword
    X:\QtS60\selfsigned.cer;X:\QtS60\selfsigned.key

To use this file the creatapackage has to be invokes as follows:

    createpackage.pl -c=mycerts.txt templatepkg release-armv5

Reviewed-by: Shane Kearns
Reviewed-by: Miikka Heikkinen