Use icons on QDialogButtonBox in KDE
Oxygen did not use icons in dialogs. This is obviously
incorrect compared with other KDE apps. We changed common style
to enable button icons by default on X11 and explicitly
disabled them for Windows and Motif styles.

We also updated the icons for YES and NO actions to use
dialog-ok and process-stop to reflect the usaged in
KDE itself.

Reviewed-by: ogoffart