exception safety fix for QList::operator+= (const QList&)
The refactoring of current++ and src++ out of the new line makes the
code easier to understand but it also seems to be significant at least
in the ::isComplex case. I suspect that the ordering increment
operations vs throw from new is not well defined, or not implemented as
you might hope (with the ++ happening very last).

The changes in the catch blocks mean that it deletes the created
objects, rather than trying with the first failed object.

The test code has been updated with a +=(Container) test, and to force
testing of both static and moveable types.

Reviewed-by: Harald Fernengel