Clean up shader programs properly
QGLShader and QGLShaderProgram used to delete the GL object id
in the wrong context.  But fixing this means we must keep track of
when contexts are destroyed so that we don't try to access a
context from a shader if it goes away.  We also need to transfer
ownership from one context to another when they are shared.

This change introduces QGLSharedResourceGuard, which keeps track
of a context and a GL object identifier.  When the context goes
away, ownership is passed to a shared context.  When there are
no more shared contexts, the identifier automatically zeroes.

Reviewed-by: trustme