Fix the bactkrace in the QScript Debugger
This patch remove the 'fake' context  that appears in the debugger
backtrace when there is a break point in the global context.

This problem never appeared in the tests because the
QScriptContext::backtrace has always at least two items in the backtrace
as it needs the native 'bt' function to be called.

Changed QScriptEnginePrivate::contextForFrame to skip the fake frame
(instead of QScriptContext::parentContext). So we never have a QScriptContext
pointing to that frame.

The changes in QScriptContextInfo are for retreiving the right filename
information for the global context when the global context is on top.

Reviewed-by: Kent Hansen