Avoid garbled output on Windows for non-ascii-compatible text
In the Windows print engine, we try to send a text item as a raw string
of characters to the printer driver if this is possible. This is to
facilitate using PDF-printers as much as possible, allowing them to
save the text in the document so for searching etc. We can only safely
do this if all the characters in the string are ASCII-compatible, i.e.
in the 7 bit range, since this is the only part of the set which is
guaranteed to be compatible across code pages.

Task-number: 180655
Reviewed-by: Trond