Fix regression in Command link button font size
Since vista style polish sets the font on polish, the
check for WA_SetFont will succeed, hence we never set the
actual font. I replaced the check with a full resolve.

However since the resolve would clear the resolve_mask,
from the widget font it has to be restored manually
so that QPainter:setFont can resolve it later.

Task-number: QTBUG-4646
Reviewed-by: trond