Account for right bearing in QFontMetrics::boundingRect(string)
QFontMetrics::boundingRect() that takes a string needs to account for
the right bearing of the last glyph, as it is documented to be the
rectangle that contains the pixels of the text. I've added a test for
this, and fixed tst_QFontMetrics::elidedText() to use boundingRect() to
find the actual width of the text drawn (width() will return the advance
of the text, which is larger than the actual width of the pixels.)

I've also fixed a small typo in the "len" -> "ilen".

Reviewed-by: Simon Hausmann