Return the correct QGLFormat to the OpenGL1 paint engine for FBO's
QGLPaintDevice::format() was returning the context's format, not
the format of the window surface's FBO.  This caused the OpenGL1
paint engine to think that the window didn't have depth and
stencil buffers, even though the FBO most certainly did.

This change makes QGLPaintDevice::format() virtual and overrides it
in QGLFBOGLPaintDevice to return an updated format that includes
the context parameters plus the extra features that the FBO supports.

Reviewed-by: Tom Cooksey