Fix crash in gl2 paint engine on Windows
wglGetProcAddress() on Windows requires a current context. Since there
was none, the resolution of GL extensions would silently fail, the
function pointers would be null and when we tried to use them later we
would get a crash (e.g. in startup of PadNavigator.) I've added an
assert to make the cause of the crash clearer, and a makeCurrent() to
fix the crash.

Reviewed-by: Tom