Fix random selection when the order is descending.
This commit fix the random selection that appear when the sort is not
ascending. The problem is that sometimes the sort is not yet made (timer
is not yet fired) so the visible children list contains both sorted items
and non sorted items (at the end). translateVisibleLocation was buggy
assuming that the list is always sorted. We have now a dirty index that
indicate where the dirty items start. And then when the sort is made
the dirty index is reset. I have added auto-test for that and fix one
that was broken for Mac. The new version of the auto-test showed a crash
because of this broken selection.

Task-number:258751
Reviewed-by:thierry