Only reset the backend pointer after we're done with it
Since 925912ebf552417306a5bd20fd986afda6a582be QNetworkReplyImplPrivate
no longer holds its own pointer to the network cache and relies on the
backend to get it. Since the cache is used in our call to finished, we
must reset the backend only after that.

This fixes a crash I was seeing in Arora...

Reviewed-by: Markus Goetz
Reviewed-by: Peter Hartmann