Remove this line; which no longer has any effect.
This line was added to fix crashes when deleting items that had
a subFocusItem pointing to a child that was already deleted. This
bug was fixed by ebb1162f54a29baeccb71d1e283146892629518f. After
this, subFocusItem is always 0 at this point.

The original change was eb3d5a73148cd7206c6b3b6672ed47b44611f745.

Reviewed-by: TrustMe