Do not use FBO as backend for window surface in opengl engine
There are too many platform regressions when using the FBO as a backend
for the opengl engine. We need to find whether it's possible to work
around them, but for the time being, to preserve compatibility, we
only use FBOs when using the opengl2 engine.

Example of issue: Smudged text and smudged diagonal lines on Windows XP
when running with opengl engine and latest nvidia driver.

Reviewed-by: Trond