Don't use UI_TOOLS for browser demo on QWS
When cross-compiling and shadow building Makefiles are not generated for
designer and libQtUiTools. For this reason designer usually doesn't
build for people trying to compile the browser example on QWS due to
unresolved externals or weird link errors when the build tries to pick
up the system Qt libraries instead. The mechanism to fix this is already
in place for S60.