Fix RTL text alignment in QComboBox items with SH_ComboBox_Popup
Change 93ba0035f4eadfaf7217d95f18a442d418a064b8 removed truncation of
a listview item's bounding rect to the viewport as documented in the
commit log. An unwanted result was that the style option passed to the
item delegates would sometimes contain a rectangle which was larger than
the delegated area. The direct result was that a QComboBox drawn with a
style with the SH_Combobox_Popup style hint in an RTL language would
align its item text to the right edge of a rectangle which was wider
than the popup menu, and thus only part of the text would be visible.

Task-number: 260974
Reviewed-by: Olivier