Fix closing the menu of a QMenuBar with the mouse
Ammend commit 41dbc4406 and 21cf7b4c431742

We do not always get the mouseReleaseEvent (because the popups grabs the
mouse.  So set mouseDown to false in the mouseMoveEvent if no buttons are
pressed.

We also need to call setCurrentAction(0) in the case we are hovering on
the empty area with no popup open.  (in otder to remove the highlight)

The code is now much closer to the code before the two previous commit

Reviewed-by: Prasanth