Implemented the SAVE_TARGET in QClipboard on X11
Added support for the simpliest SAVE_TARGET mechanizm allowing us to
work nicely with clipboard managers, so that when Qt application that
owns the clipboard exits we delegate the content to the clipboard
manager if there is one.

The current implementation doesn't specify which targets to give to
the manager, so it will try to fetch as much as possible.  Also, right
now we do not support the TARGET_SIZES target, meaning that the
manager doesn't know how much data it is going to fetch, so it will
try to fetch everything even if it can take a lot of time.

Reviewed-by: Bradley T. Hughes