Reduce overhead of paint engine-specific pixmap filters
Engine-specific pixmap filters were being created, used, and destroyed
every time draw() was called on QPixmapColorizeFilter, QPixmapBlurFilter,
and so on.  This had a heavy performance penalty and made it difficult
for the GL paint engine to cache shaders from one request to the next.

A generic filter can request an engine-specific filter that matches its
parameters.  The engine can either create a new one or return a previously
allocated filter object.

Ownership of engine-specific pixmap filter objects is moved to the
paint engine itself.

Reviewed-by: Andrew den Exter
Reviewed-by: Michael Brasser
Reviewed-by: Michael Goddard
Reviewed-by: Sarah Smith