Improve font handling on Mac/Cocoa.
The Mac/Cocoa font database currently relies on two APIs, ATSUI (old)
and CoreText (new). These are interchangable and work on the same font
database.

Some differences do exist, in particular ATSFontFamilyGetName seems
to return the same as ATSFontGetName for some fonts - the font name,
not the family name. In any case, the old ATS code path is initializing
QFontDatabase::familes() with the font name.

This causes a naming mismatch with code that uses CoreText and gets
the proper family name, in this case in initializeDb(). The fix is
to make sure the correct family name is used by using the CoreText
API to access it.

RevBy: TrustMe