QFontMetrics::boundingRect(QString) returns incorrect width on Mac OS X
This bug appears only in Carbon port (ATSUI font engine). While finding
the character stops, this font engine relys on the RTL flags. As an
optimization, for this particular version of boundingRect(), BIDI is
ignored. So for this font engine, we will always use the correct RTL
flags.

Task-number: 213109
Done-with: Eskil
Reviewed-by: Simon Hausmann