Fix QDesktopServices::storageLocation() when registry keys are missing
We should not use the registry key when looking up the storage
location as this is not really supported by Microsoft and not
allways accessible. Instead we now use SHGetSpecialFolderPath.
Note that we avoid SHGetKnownFolderPath as it is not available
on all CE platforms though it is the reccommended function by
Microsoft.

Task-number: QTBUG-3241
Reviewed-by: prasanth