QGraphicsAnchorLayout: Added an auto test for size policy
Added an auto test for the currently supported size policies set.
This is a simple test containing a single item anchored to the layout
vertical and horizontally. The size policies are set, then the
item's minimum, preferred and maximum sizes are compared to the
expected ones, according to each policy.

Signed-off-by: Anselmo Lacerda S. de Melo <anselmo.melo@openbossa.org>