Control-drag header selection behaved wierdly.
The QItemSelectionModel::Current was not set in
QTableViewPrivate::selectColumn(). However, Control-drag selection in
QTableView behaved differently than other software such as OpenOffice's
spreadsheet.

Now the behaviour when Control-dragging is that the selection will be set to
the opposite of the selection state of the first cell. If that cell is
selected, we will deselected the cells while dragging, and conversely, if it
isn't selected, the cells will be selected.

Reviewed-by: Olivier
Task-number: QT-1435
Task-number: 191545