Improved a qt_x11_wait_for_window_manager.
In addition to waiting until the window manager maps and reparents the
window, also wait until we get initial expose event.

The window manager will most probably send us several ConfigureNotify
events, so wait until all of them are handled in addition to the Expose
event (as one might rely on another).

Also, we shouldn't wait for the ReparentNotify event if the window has
X11BypassWindowManagerHint.

Reviewed-by: Olivier Goffart