Fix glyph selection in symbol fonts that contain unicode cmap table
Some symbol fonts will contain a unicode cmap table in addition to the
microsoft symbol cmap table that maps the symbol range (private range
0xf000 - 0xf100) into the correct glyphs. This is essentially a broken
unicode table, and we should not prefer it if these conditions are true.
In the strict cases where these conditions apply, we fall back to the
symbol table instead.

Task-number: QT-2354
Done-with: Lars
Reviewed by: Lars