Fix regression after code review
The branch maintained its own set of native windows.
This was commented to be a duplication of the native window list kept
by QWidgetPrivate, and changed.
Unfortunately, the set maintained on the branch contained only window
owning controls, while the list kept by QWidgetPrivate contains all
widgets that have a CCoeControl.
Added a check for whether the control is window owning before using its
DrawableWindow() to avoid getting a null pointer.

Reviewed-by: Jason Barron