Fix the scrollTo() test function on Windows.
After the QTRY_COMPARE was added, the size of the view suddenly
became a hard requirement. However, this reqirement was not met on
windows, since the forcedSize was smaller than the minimum size allowed
for the view.

Thus, it would never have its size set to forcedSize and the
QTRY_COMPARE would fail.