Allow forcing an empty window title.
When a user sets a window title on a window to an empty string we will
not check if the value was an empty string before. This handles a
special case on X11 and Windows when we set the window title to the
application executable name by default if the user didn't provide the
title, so if the user explicitely wants to remove the windowtitle
completely and make it empty, he will be able to set it to an empty
string.

Reviewed-by: Olivier Goffart
Reviewed-by: Prasanth Ullattil