Fix a bug in QDirectFBWindowSurface::scroll
It seems that one cannot do multiple blits within a double-buffered
surface in overlapping one another without doing a Flip in between. I
find this a little odd and will check that it's the correct behavior but
this fixes bugs in scroll(). To see a case that used to produce drawing
errors check out the collidingmouse example and drag the viewport
around.

Reviewed-by: Jervey Kong <jervey.kong@nokia.com>