Remove remaining support for win32-mwc mkspec
Qt 4.6 cannot be used with the win32-mwc makespec, as the Metrowerks
compiler supplied with Symbian/S60 releases does not contain enough
support to build the Qt host tools in a Unicode version, and ANSI
support has been removed in 4.6.

Alternative mkspecs are win32-g++, requiring mingw tools to be installed,
or one of the win32-msvc.... mkspecs, requiring MS Visual Studio (Express
edition is fine)

The mwc makefile for qmake was already removed, so this just removes the
remaining support.

Reviewed-by: Jason Barron