initialize the this-register of the global context
QScriptEnginePrivate::thisForContext() relies on the this-register
of the global context to contain an invalid JSValue.
The default Register constructor (used to initialize the registers
of the global context) only invalidates its value when NDEBUG is
not defined (but we define it). Therefore, we must explicitly set
the this-register to an invalid value.

Reviewed-by: Olivier Goffart