Change QErrorMessage to use the new soft key API.
With the new softkey API it is not necessaery to call setSoftKeys
with the actions you want and then to set it to 0 to clear them. The
way to use the API is just call addAction() and this soft key will be
valid while this is the active window.

Also make sure that the text of the action is updated when
retranslateUi() is called.

Reviewed-by: Alessandro Portale