Change QDialogButtonBox to use the new soft key API.
This simplifies the implementation of the soft keys. Introduce a hash
for storing the mapping between the buttons and the actions that they
represent. This allows for easy translation of the strings when the
language changes.

Use setFixedSize instead of hide to make the buttons invisible. This
allows us to put some code in the handling of the show event and add
the actions here. This should be fairly safe since even if we get
multiple show events, it is still safe to add the same action again.

Reviewed-by: Alessandro Portale