QWizard changed to use the new soft key API.
Created an array of soft key actions that are created on the fly as
buttons are needed. This moved the allocations for the actions out
of the updateButtonStates() function which is called quite often so
allocations should happen less often now.

The signals and text should are kept in sync with the buttons.
Controlling which soft keys are currently available is done when
the buttons states are updated by calling addAction or removeAction.
The soft key framework then picks up on these changes.

Reviewed-by: Alessandro Portale