Add a new compressable event QEvent::UpdateSoftKeys.
Now that the softkey framework triggers updates on focus changed,
window activated, and action added/removed/changed the softkey bar
was updating many times resulting in flicker. Solve that by
introducing a new event type that is posted to the softkey framework.
Since we only need to update the softkeys once per event loop
iteration, the event is compressible in the event loop.

Reviewed-by: Alessandro Portale