Get menubars working again with new softkey API.
This is an unfortunate workaround to get the menu bars working again.
Previously we used the soft key role to determine what action command
we should use for the softkey. Now we only use the role to determine
the position (as we should) and ideally the command would be
determined by checking if the action has a menu() and then forcing
the command to EAknSoftkeyOptions. This would have enabled any softkey
with a menu to work, but unfortunately this would require
considerable effort in the menu implementation and time is short.

So the workaround is to abuse object name to tell the softkey system
that this key is special and should get EAknSoftkeyOptions.

Reviewed-by: Alessandro Portale