Fix QPrintDialog with no parent and no active window
PrintDlgEx requires an non-null owner handle, otherwise it will fail
with an "Invalid Handle" error. This caused code which popped up a
print dialog as the only window in the application to fail silently
and immediately return Rejected from the exec() function. To continue
support for this somewhat unusual use case, we fall back to using the
print dialog itself as the owner of the print dialog sheet if all else
fails.

Reviewed-by: Trond