Improved license header autotest.
Check the license in .h files as well as .cpp files and check some
files that were previously skipped. Instead of scanning through the
whole Qt directory for .cpp and .h files, only search through a set of
subdirectories. Accept '\r' as line break in addition to '\n' and
"\r\n". Reenable the copyright check that was commented out.

Reviewed-by: Frans Englich