Fix printing with glyph substitution
If a font has substituted any of the glyphs in the string, then we
cannot convert it to a string, as the glyph positions will no longer
match the positions of the characters. E.g. in the case of the
MS Calibri font, two subsequent t's are represented by a single glyph.
Printing text in Calibri with subsequent t's will thus cause the
first t to be position and sized based on the glyph representing "tt"
and glyphs after this to be position based on the wrong advance and the
text will look mangled, with some glyphs overlapping.

Task-number: QTBUG-4445
Reviewed-by: Gunnar