Ensure that the menu only shows when clicking on the menu button
In Qt 3 clicking on the toolbutton would not popup the menu, only when
clicking on the arrow for the menu popup would this appear if the
toolbutton had MenuPopup mode set.

When a Q3ActionGroup is used and added to a toolbar then it will give a
toolbutton with such a button.  This patch fixes the behaviour so that
only clicking on the arrow button will cause the menu to appear in this
mode.

Reviewed-by: Thierry