QNAM: Avoid using toLower for case insensitive comparision
Using qstricmp is faster on the desktop than the alloc of
QByteArray::toLower and the QByteArray::operator==.

QChar::toLower will now be called for the input key from
within the loop. In the measurements done this is still
a performance improvement over the allocation done by
QByteArray::toLower. For more header values this might
not be the case anymore.

RESULT : tst_Loading::byteArrayTestLower():
     0.00076 msec per iteration (total: 25, iterations: 32768)
RESULT : tst_Loading::byteArrayTestCompare():
     0.00028 msec per iteration (total: 37, iterations: 131072)

Reviewed-By: Markus Goetz