QNAM: Remove some more occurences of toLower on the hot path
toLower will force a memory allocation, and memory allocations
are certainly adding a delay. Use qstricmp were this is easily
possible. This change makes headerValues et all drop out of
the profile.

QChar::toLower will now be called more often, specially for
loops that do linear searching. In the profile with real web
content this is faster than using QByteArry::toLower in the
inner loop and the forced memory allocations.

Reviewed-By: Markus Goetz