Fix broken drop-down arrows on Windows 7
This issue affects both tool buttons and buttosn
with menus. The problem is that the Windows metric
we were using to calculate the size changed on
Windows 7. It seems not to be the correct metric so it
seems more reliable to use the dpi-scaled constant 12 as we
were doing in common style already. This should make appearance
consistent between the windows versions.

Task-number: QTBUG-4241
Reviewed-by: denis
(cherry picked from commit 1053d8564ca33e18a813e0f0fa1ec1f381e4a937)