Fix breaking on fixed column width when text has tabs
When a text in QTextLayout contained tabs, these would be counted as
single glyphs when breaking the text on a fixed number of columns,
rather than the number of characters they span.

The patch calculates the number of characters represented by a tab by
using the average character width of the font engine.

Task-number: QTBUG-4468
Reviewed-by: Simon Hausmann